bpi-app-api/bpi-app-api/src/main/java/com/example/bpiappapi/nomenclature/service/impl/CountryServiceImpl.java
package com.example.bpiappapi.nomenclature.service.impl;

import com.example.bpiappapi.nomenclature.model.Country;
import com.example.bpiappapi.nomenclature.repository.CountryRepository;
import com.example.bpiappapi.nomenclature.service.CountryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service

public class CountryServiceImpl implements CountryService {
    private final CountryRepository countryRepository;

    @Autowired
    public CountryServiceImpl(CountryRepository countryRepository) {
        this.countryRepository = countryRepository;
    }

    @Override
    public List<Country> getAllCountries() {
        return countryRepository.findAll();

    }

    @Override
    public void saveAllCountries(List<Country> countries) {
        countryRepository.saveAll(countries);
    }
}
